+++ foundation-runtimes/terminal/console/src/commonMain/kotlin/terminal/ConsoleTerminalOutput.kt
@@ -1,0 +1,8 @@
+package terminal
+
+class ConsoleTerminalOutput(
+    config: TerminalHistoryConfig = TerminalHistoryConfig()
+) : TerminalOutput {
+    override val history = TerminalHistory(config)
+    override fun executePrint(any: Any?) = kotlin.io.println(any)
+}