+++ foundation-runtimes/terminal/console/src/jsMain/kotlin/terminal/npm/promptsync/PromptSync.kt
@@ -1,0 +1,5 @@
+package terminal.npm.promptsync
+
+import platform.utils.common.require
+
+//internal val prompt = require<() -> ((String) -> String)>("prompt-sync")()